+++ doc/version_history.md
@@ -1,14 +1,20 @@
-### Version History
+## Version History
 
+-----------------------------------------------------------------------
 
-#### 0.3.0
+##### 0.3.2
 
-+ Date: 2014-09-01
++ Date: 2014-09-17
 
-+ Add webelement/WebBrowser, for all browser operations, such as scroll to, etc.
++ Added wingui.py to support upload/download dialog on MS Windows System. (Need install "SendKeys" and "PyWin32" manually)
 
-+ Finished all normal supports. First version for real test.
++ Added support for multi-window operations.
 
++ Updated alert handling.
+
++ Fix bugs of datadriver and webelement operations.
+
+-----------------------------------------------------------------------
 
 ##### 0.3.1
 
@@ -16,6 +22,15 @@
 
 + Bug fix
 
+-----------------------------------------------------------------------
+
+#### 0.3.0
+
++ Date: 2014-09-01
+
++ Add webelement::WebBrowser, for all browser operations, such as scroll to, etc.
+
++ Finished all normal supports. First version for real test.